+++ lib/native_admob_flutter.dart
@@ -1,6 +1,6 @@
 export 'src/events.dart';
 export 'src/mobile_ads.dart';
-export 'src/utils.dart' show AdError, LoadShowAd, UniqueKeyMixin;
+export 'src/utils.dart' show AdError, LoadShowAd, UniqueKeyMixin, AdBuilder;
 
 export 'src/native/native_ad_widget.dart';
 export 'src/native/controller/controller.dart';